+++ install.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 echo "============================================="
-echo " Ampdeck+ v2.0.0 - Stream Deck Plugin"
+echo " Ampdeck+ v2.0.1 - Stream Deck Plugin"
 echo " The Unofficial Plexamp Controller"
 echo "=============================================" 
 echo ""